move DesensitizedConverter to package common
