+++ framework/src/main/resources/logback.xml
@@ -5,7 +5,7 @@
   <!-- Be sure to flush latest logs on exit -->
   <shutdownHook class="org.tron.core.config.TronLogShutdownHook"/>
   <conversionRule conversionWord="m"
-    converterClass="org.tron.common.logsfilter.DesensitizedConverter"/>
+    converterClass="org.tron.common.log.layout.DesensitizedConverter"/>
 
   <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
     <encoder>
